+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [MODULE] Added new browser support: Lynx and W3m
 ### 0.19.1
 - [Firefox based] Fix an error when `load()` fails if librewolf or firefox is not installed
 ### 0.19.0